jarvis_mcp/requirements.txt
mcp>=0.1.0
ollama>=0.1.0
requests>=2.31.0
psutil>=5.9.0
Pillow>=10.0.0
